Move Task function to logic.js
