+++ src/logic.js
@@ -77,9 +77,19 @@ const projectInputVerification = (projectTitle) => {
     }
 }
 
+function Task(title, description, dueDate, priority, notes, checklist) {
+    this.title = title;
+    this.description = description;
+    this.dueDate = dueDate;
+    this.priority = priority;
+    this.notes = notes;
+    this.checklist = checklist;
+}
+
 export { createE }
 export { setActiveButton };
 export { showCalledProject };
 export { deleteOldProjectsFromHTML };
 export { loopThroughLocalStorageKeys };
 export { projectInputVerification };
+export { Task };
